New recompilation scheme.
The original scheme was just recompile the changed files, if the linking
phase fails with "Inconsistent assumptions ..." error, gather the
mentioned files and recompile them and try linking again.

The problem is that this does not work for if one those offenders is in
a library one level below the compiled modules.

In such a case the build entered an infinite loop and the only to stop
it was to quit the editor.

New scheme is probably prone to recompilation, but I can live with that.