Fixiing the plugin-diff-native build.
The heuristic if it writes to stderr it failed, failed again.

The plugin need to compile 2 files before and it complains about missing
.cmx file and and .cmi being not compiled with -opaque.

But it compiles. The first file. Then the build process is aborted, so
the second file is never compiled. And the build fails.

As for the good news the build.ml diffs are now readable. Hurray.