Change the compilation scheme for autocomp
In the previous scheme, the `ocamlc` compiler was invoked from the `src`
directory and was compiling a file copied to `.tmp` directory

The advantages of this method is that we could take the include paths
from the project as is.

The disadvantages is that the locations in `foo.cmi` file do not point to
`foo.ml` but to `../.tmp/foo.ml`. This changes the CRC of module
interface even if the module stayed the same.

If we copy the fresh `.cmi` over the old one modules which depend on it
complain about incompatible interfaces, and if we do not copy it, we
need a full rebuild to use the changed module.

The copying was removed, as part of a fix for infinite oebuild loop, but
I hope to add it back in the next commit.