Remove warning -48 from autocomp options
This is the one that warns you that one or several optional parameters are missing.
It is allmost always a false positive and what you intend anyway.

I was motivated by this article: https://dev.to/yawaramin/ocaml-reasonml-best-practice-warnings-and-errors-4mkm

I'm not turning off the -42 warning not because compiling with an OCaml compiler is valid use case for me,
but because I found it more readable to have explicit module name in more places than less.