Some binannot modules refactoring.
No functionality changes. Just removing the `open`-ing of binannot
modules and copying the types all over.

Since OCaml type are not nominal, we can as well use it.
Is just a matter of personal preference I guess.

Anyway, now the type definitions are closer to their use, and function
invokation are prefix by the module in which those are defined.