Swith to compiler-libs Tast_iterator based .cmt processing
The old mechanism of extracing information from the `.cmt` files was
relying on hand-written `iter_pattern`, `iter_expression`.. functions
for traversing the the typedtree intermediate representation of OCaml
code.

For the moment the old code is retained, but is disabled by default.
A new iteration scheme based on `Tast_iterator` was added. The idea is
to reuse the existing mechanism which already in place in compiler-libs
and which should, I hope, make the adaption to future versions of OCaml
easier.

To enable the old mechanism one should set the environment variable
`OEANNOTBIN=v1`