Simplest preprocessor for ocamldoc
Since the switch from `Pervasives` to `Stdlib` you
cannot just call `ocamldoc` on `stdlib.{ml,mli}`

It still refers to `Arg`, `String` modules, etc. but
but `.cmi` files are `stdlib__arg.cmi`, `stdlib__string.cmi`
and so on.

So we have to invoke the `ocamldoc -pp stdlib_pp ..`
to fix that.

This is just the preprocessor. Changes to the Ocamleditor
itself will be in the next commit