Switch to ocp-indent.lib for formatting
Instead of shelling out to ocp-indent, now the ocp-indent.lib
is used to calculate the required offsets.

It means that ocp-indent is no longer optional.

TODO1: Use the `IndentConfig.man` to extract and display
the ocp-indent help.

TODO2: Currently you can have different indentation for
edtior and `ocp-indent`. This need to be fixed somehow.

All will be good if you keep the same indentation for both.
By hand.