Do not indent non OCaml files.
This fixes the minor but embarrassing bug, where
the editor when the 'format on save' box is checked will
try to indent as OCaml files all files.

The result can be surprising, especially if your file
contains some OCaml keywords.

As a small extra, I've added support for indentation
of ocamllex `.mll`, files. For the moment, just on save.

Also we need to disable the Indent/Indent All for files
for which this is not supported.

But that's another issue.