Refactoring and idle mode functionality
In this commit, code refactoring has been performed, and an idle mode
functionality has been introduced. It was observed that the enter_notify
event did not function as expected on the text view. Therefore, a custom
solution has been implemented to address situations where the pointer exits
the window without losing focus. The idle mode utilizes a low-speed timer
to detect the pointer's entry into the text view.