Change the undo behavior: Undo all contiguous inserts if recorded within a few milliseconds of each other. Makes the buffer readonly from the time merlin is invoked to get folding point offsets until the offsets are processed.
This is to avoid crashes due to the processing of offsets that have
become invalid during the processing itself. This could occur when
typing very quickly. The same problem also occurred during undo
operations (e.g. holding down Ctrl+Z for a long time). In this case the
workaround is to undo all the contiguous insertions recorded at a very
short distance from each other.