Enable -bin-annot-occurrences flag
This commit adds the -bin-annot-occurrences flag to the build configurations in build.ml, oebuild.ml, and project.ml to enable the generation of occurrence annotations, which are used by Merlin for symbol lookup.

Additionally, this commit includes the following changes:

Refactor: The Merlin.occurrences function is updated to support project-wide searches by generating an index file using ocaml-index.

Chore: The *.ocaml-index file is added to .gitignore.